# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJN Luxury Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin or customer own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track/{trackingCode}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by tracking code (public)
      parameters:
        - name: trackingCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"
  /orders/track-by-phone/{last4}:
    get:
      operationId: trackOrdersByPhone
      tags: [orders]
      summary: Find orders by last 4 digits of phone (public)
      parameters:
        - name: last4
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching orders (latest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderTracking"
  /orders/my:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: Get current customer orders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: My orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # ─── GALLERY ─────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Customer token returned by /auth/verify-otp.
    AdminCookieAuth:
      type: apiKey
      in: cookie
      name: ajn_admin_session
      description: HttpOnly admin session cookie returned by /admin/auth/login.
...
  schemas:
...
    Order:
      type: object
      required: [id, trackingCode, status, total, createdAt]
      properties:
        id:
          type: integer
        trackingCode:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        status:
          type: string
        serviceType:
          type: ["string", "null"]
        total:
          type: number
        deliveryFee:
          type: number
        paymentMethod:
          type: string
          enum: [cod, transfer, paid]
        governorate:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        mapsUrl:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: string
        quantity:
          type: integer
        price:
          type: number
        selectedColor:
          type: ["string", "null"]
        customization:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [customerName, customerPhone, governorate]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        governorate:
          type: string
        area:
          type: string
        address:
          type: string
        notes:
          type: string
        paymentMethod:
          type: string
          enum: [cod, transfer, paid]
        deliveryZoneId:
          type: integer
        mapsUrl:
          type: string
          description: Google Maps URL (auto-generated from geolocation or manually entered)
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string
    OrderTracking:
      type: object
      required: [trackingCode, status, createdAt]
      properties:
        trackingCode:
          type: string
        status:
          type: string
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        kind:
          type: string
          description: "product | service"
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
        createdAt:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
        mapsUrl:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
        eventLocation:
          type: ["string", "null"]
        customerConfirmation:
          type: ["string", "null"]
          description: "confirmed | reschedule_requested | null"
        requestedDate:
          type: ["string", "null"]
        confirmationNote:
          type: ["string", "null"]
        confirmationAt:
          type: ["string", "null"]
...
    StatusHistoryEntry:
      type: object
      required: [status, createdAt]
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── GALLERY ──────────────────────
...